Improved and refactoring UI API models
