+++ client/api/index.ts
@@ -1,9 +1,12 @@
+import { APIMeteo } from '@/api/apiMeteo'
+
 import * as ApiModel from './models'
 import * as ApiType from './types'
 import { API, HOST_API, HOST_IMG, SITE_LINK } from './api'
 import { useAppDispatch, useAppSelector } from './store'
 
 export {
+    APIMeteo,
     ApiModel,
     ApiType,
     API,
