+++ client/api/models/weather.ts
@@ -1,14 +1,17 @@
 export type Weather = {
-    condition_id: number
-    temperature: number
-    temperature_feels: number
-    humidity: number
-    pressure: number
-    wind_speed: number
-    wind_gust: number
-    wind_degree: number
-    clouds: number
-    precipitation: number
-    illumination: number
-    uvindex: number
+    temperature?: number
+    feelsLike?: number
+    pressure?: number
+    humidity?: number
+    dewPoint?: number
+    visibility?: number
+    uvIndex?: number
+    solEnergy?: number
+    solRadiation?: number
+    clouds?: number
+    precipitation?: number
+    windSpeed?: number
+    windGust?: number
+    windDeg?: number
+    date?: string
 }
