+++ client/api/types/weather.ts
@@ -1,20 +1,9 @@
 import { ApiModel } from '@/api'
 
-/* Current */
-export interface ResCurrent {
-    timestamp: {
-        server: number
-        update: number
-    }
-    conditions: ApiModel.Weather
-}
-
 /* Statistic */
-export interface ReqStatistic {
-    period?: string
+export interface RequestHistory {
+    start_date: string
+    end_date: string
 }
 
-export interface ResStatistic {
-    date: string
-    weather: ApiModel.Weather[]
-}
+export type ResponseHistory = ApiModel.Weather[]
